2.5 Создал модули (отдельные файлы) и перенес в них написанные в прошлом задании функции для генерации DOM-элементов из main.js
