+++ src/components/trip-routes.js
@@ -1,0 +1,7 @@
+export const createTripRoutesContainerTemplate = (child) => {
+  return (
+    `<ul class="trip-days">
+      ${child}
+    </ul>`
+  );
+};
