+++ src/components/trip-tabs.js
@@ -1,0 +1,9 @@
+export const createTripTabsTemplate = () => {
+  return (
+    `<h2 class="visually-hidden">Switch trip view</h2>
+    <nav class="trip-controls__trip-tabs  trip-tabs">
+      <a class="trip-tabs__btn" href="#">Table</a>
+      <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Stats</a>
+    </nav>`
+  );
+};
